Problem 1 and Problem 2
Problem 1
You were playing a fun guessing game during your school break. There were a total of 100 participants excluding you. Out of these 100 people, 30 were Maths Majors, 45 were Economics Majors and 25 were Physics Majors.

The game was divided into three rounds.

In the first round, you had to guess the number of Maths Majors and you correctly guessed 20 of them.
In the second round, you had to guess the number of Economics Majors and you correctly guessed 30 of them.
In the final third round, you had to guess the number of Physics Majors and you correctly guessed 20 of them.
Accuracy is defined as the number of correct guesses upon total number of people in the group (expressed in percentage)


Problem 2
A bag contains 45 apples, 65 oranges and 30 bananas. Find the percentage of each type of food items in the bag.